Move constants related to actions to be inside actions folder
