+++ src/actions/query.js
@@ -6,7 +6,7 @@ import {
   EXECUTE_QUERY_SUCCESS,
   EXECUTE_QUERY_FAILURE,
   UPDATE_SQL_SUCCESS
-} from '../constants/action-types';
+} from './types';
 
 
 export function executeQuery(sql) {
