+++ src/reducers/connections.js
@@ -1,4 +1,4 @@
-import { LOAD_CONNECTIONS_SUCCESS } from '../constants/action-types';
+import { LOAD_CONNECTIONS_SUCCESS } from '../actions/types';
 
 const initialState = [];
